[DAPK-17]~bug fix for model initialization, uploading and removing
