+++ 0_run_upload_update.bat
@@ -1,3 +1,3 @@
 
-.\python_embeded\python.exe -s Fooocus\entry_update.py  --preset default%*
+.\python_embeded\python.exe -s Fooocus\entry_update.py  --preset default %*
 pause
