+++ Fooocus/launch_update.py
@@ -20,7 +20,7 @@ import fooocus_version
 
 from build_launcher import build_launcher
 from modules.launch_util import is_installed, run, python, run_pip, requirements_met, delete_folder_content
-from modules.model_loader import load_file_from_url
+from modules.model_loader_util import ini_args, download_models, upload_add_files
 
 REINSTALL_ALL = False
 TRY_INSTALL_XFORMERS = False
@@ -67,9 +67,9 @@ vae_approx_filenames = [
 ]
 
 
-def ini_args():
-    from args_manager import args
-    return args
+# def ini_args():
+#     from args_manager import args
+#     return args
 
 
 prepare_environment()
@@ -101,47 +101,53 @@ if config.temp_path_cleanup_on_launch:
         print(f"[Cleanup] Failed to delete content of temp dir.")
 
 
-def download_models(default_model, previous_default_models, checkpoint_downloads, embeddings_downloads, lora_downloads, vae_downloads):
-    for file_name, url in vae_approx_filenames:
-        load_file_from_url(url=url, model_dir=config.path_vae_approx, file_name=file_name)
+# def download_models(default_model, previous_default_models, checkpoint_downloads, embeddings_downloads, lora_downloads, vae_downloads):
+#     for file_name, url in vae_approx_filenames:
+#         load_file_from_url(url=url, model_dir=config.path_vae_approx, file_name=file_name)
 
-    load_file_from_url(
-        url='https://huggingface.co/lllyasviel/misc/resolve/main/fooocus_expansion.bin',
-        model_dir=config.path_fooocus_expansion,
-        file_name='pytorch_model.bin'
-    )
+#     load_file_from_url(
+#         url='https://huggingface.co/lllyasviel/misc/resolve/main/fooocus_expansion.bin',
+#         model_dir=config.path_fooocus_expansion,
+#         file_name='pytorch_model.bin'
+#     )
 
-    if args.disable_preset_download:
-        print('Skipped model download.')
-        return default_model, checkpoint_downloads
+#     if args.disable_preset_download:
+#         print('Skipped model download.')
+#         return default_model, checkpoint_downloads
 
-    if not args.always_download_new_model:
-        if not os.path.exists(os.path.join(config.paths_checkpoints[0], default_model)):
-            for alternative_model_name in previous_default_models:
-                if os.path.exists(os.path.join(config.paths_checkpoints[0], alternative_model_name)):
-                    print(f'You do not have [{default_model}] but you have [{alternative_model_name}].')
-                    print(f'Fooocus will use [{alternative_model_name}] to avoid downloading new models, '
-                          f'but you are not using the latest models.')
-                    print('Use --always-download-new-model to avoid fallback and always get new models.')
-                    checkpoint_downloads = {}
-                    default_model = alternative_model_name
-                    break
+#     if args.always_download_new_model:
+#         if os.path.exists(os.path.join(config.paths_checkpoints[0], default_model)):
+#             for alternative_model_name in previous_default_models:
+#                 if os.path.exists(os.path.join(config.paths_checkpoints[0], alternative_model_name)):
+#                     print(f'You do not have [{default_model}] but you have [{alternative_model_name}].')
+#                     print(f'Fooocus will use [{alternative_model_name}] to avoid downloading new models, '
+#                           f'but you are not using the latest models.')
+#                     print('Use --always-download-new-model to avoid fallback and always get new models.')
+#                     checkpoint_downloads = {}
+#                     default_model = alternative_model_name
+#                     break
 
-    for file_name, url in checkpoint_downloads.items():
-        load_file_from_url(url=url, model_dir=config.paths_checkpoints[0], file_name=file_name)
-    for file_name, url in embeddings_downloads.items():
-        load_file_from_url(url=url, model_dir=config.path_embeddings, file_name=file_name)
-    for file_name, url in lora_downloads.items():
-        load_file_from_url(url=url, model_dir=config.paths_loras[0], file_name=file_name)
-    for file_name, url in vae_downloads.items():
-        load_file_from_url(url=url, model_dir=config.path_vae, file_name=file_name)
+#     for file_name, url in checkpoint_downloads.items():
+#         load_file_from_url(url=url, model_dir=config.paths_checkpoints[0], file_name=file_name)
+#     for file_name, url in embeddings_downloads.items():
+#         load_file_from_url(url=url, model_dir=config.path_embeddings, file_name=file_name)
+#     for file_name, url in lora_downloads.items():
+#         load_file_from_url(url=url, model_dir=config.paths_loras[0], file_name=file_name)
+#     for file_name, url in vae_downloads.items():
+#         load_file_from_url(url=url, model_dir=config.path_vae, file_name=file_name)
 
-    return default_model, checkpoint_downloads
+#     return default_model, checkpoint_downloads
+
+@upload_add_files
+def initial_download_models(*params, **kwparams):
+    print('[INFO] Preparing environment to initial models uploading')
+    return  download_models(*params, **kwparams)
 
 
-config.default_base_model_name, config.checkpoint_downloads = download_models(
+config.default_base_model_name, config.checkpoint_downloads = initial_download_models(
+    vae_approx_filenames,
     config.default_base_model_name, config.previous_default_models, config.checkpoint_downloads,
-    config.embeddings_downloads, config.lora_downloads, config.vae_downloads)
+    config.embeddings_downloads, config.lora_downloads, config.vae_downloads, args)
 
 config.update_files()
 init_cache(config.model_filenames, config.paths_checkpoints, config.lora_filenames, config.paths_loras)
